chore: add setup/shutdown-validators command
