+++ helpers/hardhat-constants.ts
@@ -19,6 +19,7 @@ const getPrivateKeyfromEncryptedJson = (
     : "";
 
 export const HARDHAT_CHAINID = 31337;
+export const ZHEJIANG_CHAINID = 1337803;
 export const GOERLI_CHAINID = 5;
 export const FORK_CHAINID = 522;
 export const MAINNET_CHAINID = 1;
